feat: SearchDialog 내부 control 패널들 분리
