+++ src/features/Search/CreditsControl.tsx
@@ -1,0 +1,35 @@
+import { FormControl, FormLabel, Select } from "@chakra-ui/react";
+import { SearchOption } from "./SearchDialog";
+import { memo } from "react";
+
+interface CreditsControlProps {
+  credits: SearchOption["credits"];
+  changeSearchOption: (
+    field: keyof SearchOption,
+    value: SearchOption[typeof field]
+  ) => void;
+}
+
+const CreditsControl = ({
+  credits,
+  changeSearchOption,
+}: CreditsControlProps) => {
+  return (
+    <FormControl>
+      <FormLabel>학점</FormLabel>
+      <Select
+        value={credits}
+        onChange={(e) => changeSearchOption("credits", e.target.value)}
+      >
+        <option value="">전체</option>
+        <option value="1">1학점</option>
+        <option value="2">2학점</option>
+        <option value="3">3학점</option>
+      </Select>
+    </FormControl>
+  );
+};
+
+CreditsControl.displayName = "CreditsControl";
+
+export default memo(CreditsControl);
