+++ src/features/Search/GradesControl.tsx
@@ -1,0 +1,41 @@
+import {
+  Checkbox,
+  CheckboxGroup,
+  FormControl,
+  FormLabel,
+  HStack,
+} from "@chakra-ui/react";
+import { SearchOption } from "./SearchDialog";
+import { memo } from "react";
+
+interface GradesControlProps {
+  grades: SearchOption["grades"];
+  changeSearchOption: (
+    field: keyof SearchOption,
+    value: SearchOption[typeof field]
+  ) => void;
+}
+
+const GradesControl = ({ grades, changeSearchOption }: GradesControlProps) => {
+  return (
+    <FormControl>
+      <FormLabel>학년</FormLabel>
+      <CheckboxGroup
+        value={grades}
+        onChange={(value) => changeSearchOption("grades", value.map(Number))}
+      >
+        <HStack spacing={4}>
+          {[1, 2, 3, 4].map((grade) => (
+            <Checkbox key={grade} value={grade}>
+              {grade}학년
+            </Checkbox>
+          ))}
+        </HStack>
+      </CheckboxGroup>
+    </FormControl>
+  );
+};
+
+GradesControl.displayName = "GradesControl";
+
+export default memo(GradesControl);
